backend/routes/ai_upload.py
from fastapi import APIRouter, File, UploadFile, Form, HTTPException
from db_client import supabase
from PIL import Image
import imagehash
import io
import json

router = APIRouter(prefix="/ai", tags=["AI Verification"])

@router.post("/verify-upload")
async def verify_upload(
    ngo_id: str = Form(...),
    project_description: str = Form(""),
    file: UploadFile = File(...)
):
    try:
        # 1. Image hashing for duplicate detection
        contents = await file.read()
        img = Image.open(io.BytesIO(contents))
        phash = str(imagehash.phash(img))
        
        duplicate_flag = False
        duplicate_ngos = []
        score = 85
        verdict = "Authentic field proof detected. Content matches stated NGO objectives."
        
        # 2. Check for duplicate fingerprints in Supabase
        # Note: imagehash comparison is ideally done via Hamming distance, but for this project we compare exact hash
        res = supabase.table("proof_fingerprints").select("ngo_id").eq("image_hash", phash).execute()
        
        if res.data:
            duplicate_flag = True
            score = 30
            verdict = "AI ALERT: Possible duplicate or reused proof detected in global database. Further audit required."
            duplicate_ngos = list(set([row["ngo_id"] for row in res.data if row["ngo_id"] != ngo_id]))

        # 3. Label generation (simple keyword mapping for hackathon)
        labels = ["impact", "field-proof"]
        desc_lower = project_description.lower()
        
        if any(w in desc_lower for w in ["school", "education", "student", "teacher"]):
            labels.extend(["education", "learning-impact"])
        if any(w in desc_lower for w in ["environment", "tree", "plant", "cleanup"]):
            labels.extend(["environment", "eco-transparency"])
        if any(w in desc_lower for w in ["health", "medical", "camp", "hospital", "doctor"]):
            labels.extend(["healthcare", "medical-camp"])
        if any(w in desc_lower for w in ["food", "meal", "hunger", "nutrition"]):
            labels.extend(["nutrition", "food-security"])

        # 4. Final metadata results
        return {
            "score": score,
            "verdict": verdict,
            "labels": list(set(labels)),
            "duplicate_flag": duplicate_flag,
            "duplicate_ngos": duplicate_ngos,
            "image_hash": phash
        }

    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))
