backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pyjwt
httpx
pillow
imagehash
google-genai
python-multipart
anthropic
